Removed the `UserSession` model.
The `UserSession` model will not be used within this new version of the
application. In this new version, we will always use APITokens, for
everything, including interactive sessions. A interactive session will
be a APIToken without a APIClient.

Adding APIClient as a relationship to APITokens will be fixed in issue
#29.

This commit fixes #30.