Updated `SQLmodel` and `Pydantic`
This gives an issue; the `Field` object of `sqlmode` is not updated to
reflect the arguments of the `Field` object of the new version of
`pydantic`. This causes the `regex` parameter to not work anymore. I
fixed this temporarly by using the `Field` object from `pydantic`. I
have a raised a discussion on GitHub about this to see if something can
be done to fix this.